fix: create_comment view no longer marks seen
It is a bad idead to have the comments to be marked as seen selected by
the posting comment view in django. Rather a list with comments to be
marked as seen should be sent along the new comment itself to avoid any
race condition where a student sends a new question while staff
answered resulting in staff never seeing that new post.